refactor: deduplicate message deepFreeze helper
