+++ scripts/config/contentSafe/highlighterMessages.js
@@ -2,29 +2,7 @@
  * Highlighter-safe UI messages, supporting tree-shaking for content-injected modules.
  */
 
-const FREEZABLE_TYPES = new Set(['object', 'function']);
-
-function isFreezableValue(value) {
-  return value !== null && FREEZABLE_TYPES.has(typeof value);
-}
-
-function deepFreeze(target) {
-  if (!isFreezableValue(target)) {
-    return target;
-  }
-
-  if (Object.isFrozen(target)) {
-    return target;
-  }
-
-  for (const value of Object.values(target)) {
-    if (isFreezableValue(value)) {
-      deepFreeze(value);
-    }
-  }
-
-  return Object.freeze(target);
-}
+import { deepFreeze } from '../shared/deepFreeze.js';
 
 export const HIGHLIGHTER_MESSAGES = deepFreeze({
   FLOATING_RAIL: {
