+++ scripts/config/shared/backgroundMessages.js
@@ -2,29 +2,7 @@
  * Background-safe UI response messages, supporting tree-shaking for background handlers.
  */
 
-const FREEZABLE_TYPES = new Set(['object', 'function']);
-
-function isFreezableValue(value) {
-  return value !== null && FREEZABLE_TYPES.has(typeof value);
-}
-
-function deepFreeze(target) {
-  if (!isFreezableValue(target)) {
-    return target;
-  }
-
-  if (Object.isFrozen(target)) {
-    return target;
-  }
-
-  for (const value of Object.values(target)) {
-    if (isFreezableValue(value)) {
-      deepFreeze(value);
-    }
-  }
-
-  return Object.freeze(target);
-}
+import { deepFreeze } from './deepFreeze.js';
 
 export const BACKGROUND_MESSAGES = deepFreeze({
   POPUP: {
